+++ processor.h
@@ -13,10 +13,10 @@ private:
         l1cache* Dcache;
 public:
         processor(l1cache* icache,l1cache* dcache,watcher* thewatcher);
-        int decode(char op, unsigned long long int address, unsigned int bytes);
-        int fetchInstr(ull address, uint bytes);
-        int read(ull address,uint bytes);
-        int write(ull address, uint bytes);
+        ull decode(char op, ull address, uint bytes);
+        ull fetchInstr(ull address, uint bytes);
+        ull read(ull address,uint bytes);
+        ull write(ull address, uint bytes);
 };
 
 #endif
